backend/dgbackend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-db-connection-pool==1.2.5
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
psycopg==3.1.18
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
sqlparams==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
